Manual roll of Dart b29f228f62...016e8880f0 (flutter/engine#20967)
dart-lang/sdk@016e8880f0a Version 2.10.0-90.0.dev
dart-lang/sdk@34f636ce1d8 Revert "[vm/io] Fix FileSystemWatcher enum name typo."
dart-lang/sdk@7a02118a410 [vm/fuchsia]: Safely shutdown the async message loop.
dart-lang/sdk@334d3681467 [analyzer_plugin] Remove unused dart:async imports
dart-lang/sdk@047c0f860a5 [Analyzer] Support analyzing open files without open workspaces
dart-lang/sdk@f846d0e7ea0 [analyzer_cli] Remove unused dart:async imports
dart-lang/sdk@ba372d6e8a4 [dartdev] Remove unused dart:async import
dart-lang/sdk@30cfb228c00 [nnbd_migration] Remove unused dart:async imports
dart-lang/sdk@1134a0f17c8 [analysis_server] Remove unused dart:async imports
dart-lang/sdk@3317d47c7de [_fe_analyzer_shared] Remove unused dart:async import
dart-lang/sdk@0a036db5962 [vm/io] Fix FileSystemWatcher enum name typo.
dart-lang/sdk@000f3135a5c [analyzer] Remove unused dart:async imports
dart-lang/sdk@7817aa17967 [dartfix] Remove unused dart:async imports
dart-lang/sdk@0144087619f Version 2.10.0-89.0.dev
dart-lang/sdk@42dfb9d3046 Add `generateCallGraphWithDominators` method to generate a `CallGraph` from a precompiler trace.
dart-lang/sdk@01024dfa5a8 Issue 43200. Report ASSIGNMENT_TO_FINAL when 'final late' has an initializer.
dart-lang/sdk@ab16d79af97 Revert "Flow analysis changes to fix mixed-mode unsoundness loophole."
dart-lang/sdk@7363adca67a [infra] Update checked-in SDKs to 2.10.0-79.0.dev
dart-lang/sdk@d833f2f65c1 Flow analysis changes to fix mixed-mode unsoundness loophole.
dart-lang/sdk@33e4a6bc2a8 Revert "[vm] Have Read/WriteStream use (S)LEB128 for variable-length encoding."
dart-lang/sdk@d9dc24b53e7 Version 2.10.0-88.0.dev
dart-lang/sdk@7130dd4d3d9 [cfe] Report error on non-void setter return type
dart-lang/sdk@a7cb764aa45 [tests/ffi] Reinstate deleted test
dart-lang/sdk@39689eb07cd [co19] Roll co19 to 9c07e26d67665de7972ba6f7e87369933a978545
dart-lang/sdk@1f92293fd58 Revert "[dart:html] Add 'creates' type description to JS calls"
dart-lang/sdk@e44cf969c64 Fix creates value for getStats
dart-lang/sdk@15201ebe5f1 Version 2.10.0-87.0.dev
dart-lang/sdk@eb8dfee1074 [CFE] Clean up duplicate entries in spell checker lists

+++ engine/src/flutter/lib/ui/lerp.dart
@@ -20,7 +20,7 @@ double? lerpDouble(num? a, num? b, double t) {
   assert(a.isFinite, 'Cannot interpolate between finite and non-finite values');
   assert(b.isFinite, 'Cannot interpolate between finite and non-finite values');
   assert(t.isFinite, 't must be finite when interpolating between values');
-  return a * (1.0 - t) + b * t as double;
+  return a * (1.0 - t) + b * t;
 }
 
 /// Linearly interpolate between two doubles.
